add new cards, tweak gen_cards.sh a bit

+++ scripts/gen_cards.sh
@@ -5,12 +5,16 @@
 
 if false
 then
-for i in $(adb shell su -c ls '/data/data/com.blizzard.wtcg.hearthstone/files/Data/dxt/*.unity3d')
+for i in $(adb shell su -c ls '/data/data/com.blizzard.wtcg.hearthstone/files/Data/dxt/*.unity3d' | sed 's/\r//g')
 do
   BASENAME=$(basename $i)
-  adb shell su -c cp $i /sdcard/${BASENAME}
-  adb pull /sdcard/${BASENAME}
-  adb shell rm /sdcard/${BASENAME}
+  if [[ ${#i} -gt 2 ]]  #somehow the ls might append empty chars
+  then
+    echo $BASENAME
+    adb shell su -c cp $i /sdcard/${BASENAME}
+    adb pull /sdcard/${BASENAME}
+    adb shell rm /sdcard/${BASENAME}
+  fi
 done
 fi
